added abstract classes and update docs
