+++ lib/cbdocument.utils.js
@@ -139,7 +139,7 @@ exports.buildRefValues = function (indexes, data) {
       if (typeof value === 'number') {
         valToStore = value.toString();
       }
-      if (typeof value === 'object' && value._isBaseObject) {
+      if (typeof value === 'object' && value._isBaseModel) {
         valToStore = value.getDocumentKeyValue(value.config.storeFullReferenceId);
       }
       else if (Array.isArray(value)) {
@@ -151,7 +151,7 @@ exports.buildRefValues = function (indexes, data) {
             if (typeof v === 'number') {
               return v.toString();
             }
-            if (typeof v === 'object' && v._isBaseObject) {
+            if (typeof v === 'object' && v._isBaseModel) {
               return v.getDocumentKeyValue(v.config.storeFullReferenceId);
             }
           }
