+++ lib/document.js
@@ -1,10 +1,10 @@
 const _ = require('lodash');
 const uuid = require('uuid');
 
-import BaseObject from './baseobject';
+import { BaseModel } from './basemodel';
 import _privateKey from './privatekey';
 
-export default class Document extends BaseObject {
+export default class Document extends BaseModel {
   /**
    * @classdesc Base constructor for all created Document instances.
    * Represents just the document data and generic properties and functions.
@@ -12,7 +12,7 @@ export default class Document extends BaseObject {
    *
    * @description Clients do not need to create Document manually.
    * @class
-   * @augments BaseObject
+   * @augments BaseModel
    * @param {Object} values - the object data
    * @param {Object} options - creation options
    * @param {Boolean} options.clone - Whether to deep clone the incoming data. Default: <code>false</code>.
@@ -41,9 +41,6 @@ export default class Document extends BaseObject {
     if (this.init && _.isFunction(this.init)) {
       this.init();
     }
-
-    // May return actual object instance or Proxy, depending on harmony support.
-    return this[_privateKey]._this;
   }
 
   /**
