+++ lib/index.js
@@ -4,7 +4,7 @@ if (!global._babelPolyfill) {
   require('babel-polyfill');
 }
 
-// Without the --harmony and --harmony_proxies flags, options strict: false and dotNotation: true will fail with exception
+// Without the --harmony and --harmony-proxies flags, options strict: false and dotNotation: true will fail with exception
 if (typeof Proxy !== 'undefined') {
   require('harmony-reflect');
 }
