+++ lib/model.js
@@ -3,7 +3,7 @@ const _ = require('lodash');
 const hooks = require('hooks-fixed');
 const inflection = require('inflection');
 
-import BaseObject from './baseobject';
+import { PlainBaseModel } from './basemodel';
 import Schema from './schema';
 import CouchbaseDocument from './cbdocument';
 import _privateKey from './privatekey';
@@ -23,31 +23,31 @@ export class Model extends CouchbaseDocument {}
  */
 
 function compilePlainObject(schema, options) {
-  class ObjectInstance extends BaseObject {
+  class PlainModelInstance extends PlainBaseModel {
     constructor(data, options = {}) {
       super(data, options, schema);
     }
   }
 
-  ObjectInstance.schema = schema;
+  PlainModelInstance.schema = schema;
 
   // if the user wants to allow modifications
   if (options.freeze !== false) {
-    Object.freeze(ObjectInstance);
+    Object.freeze(PlainModelInstance);
   }
 
-  return ObjectInstance;
+  return PlainModelInstance;
 }
 
 /**
- * Compiles a descriptor into an <code>ObjectInstance</code> or <code>ModelInstance</code>
+ * Compiles a descriptor into an <code>PlainModelInstance</code> or <code>ModelInstance</code>
  * @param {Schema|Object} descriptor - the schema or schema descriptor. If <code>Object</code> we create a new schema and
- *                                     create an <code>ObjectInstance</code>, otherwise we create a <code>ModelInstance</code>
+ *                                     create an <code>PlainModelInstance</code>, otherwise we create a <code>ModelInstance</code>
  * @param {Object} options - Schema creation options
  * @param {String|undefined} name - the name of the model
  * @param {Driver|null} db - the database driver
  * @param {Object} config - the config
- * @returns {ObjectInstance|ModelInstance}
+ * @returns {PlainModelInstance|ModelInstance}
  * @private
  */
 export function compile(descriptor, options = {}, name = undefined, lounge) {
